EQC: Weaken the run/2 post condition.
The problem with the `run/2` post condition is that it is racy. The code basically does

	ask
	if ok:
		run
		record eventual melt

The problem is that two processes might both ask and then get ok and run. But the first of these will trigger a melt which the other will see. Since state transition in the code is atomic, this leads to an impossible interleaving where both manages to succeed through the code.