Simplify time-stamping code, optimize run/3 path.
Two changes, the latter first:

The call `run/3` is moved from module `fuse` into module `fuse_srv` since it is gen_server local. It was odd
to have all the functions but `run/3` "call-through" to the `fuse_srv`, so now it is inside for completeness.

The time-stamping code is updated so we time-stamp inside the gen_server, and avoid pushing in time from
the outside. The idea is to *mock* the timestamp under test so we can control when time passes by elapsing
it from the model in question. Thomas Arts has provided a small timer construction which can play the role
as a timer system.

The model still needs updating to reflect this new world.