Do not increase time if we ask for a timestamp in `fuse_time`.
This is somewhat debatable. If we use `erlang:now/0` we are guaranteed a time-increase whenever we ask for the timestamp. However, if we use `os:timestamp()` it is possible, though highly unlikely, that multiple processes will end up with the same timestamp in the `fuse_srv`. In principle, a very very fast machine would be able to process a melt in less than a microsecond, which gives the trouble.