EQC: Fix model to handle timeouts correctly.
Observation: we only use time tracking for one thing in the model: to track if a melt has expired in the time window. Once the fuse blows, we don't use time for anything anymore. So we can simplify the model a lot. A fuse can always be reset due to a timeout if it is blown. We don't need to track this explicitly. Some timeouts will arrive too soon and some will arrive too late. But the code ought to be robust even when this happens.

So we remove the notion of time from the timeout resets. This seems to solve the current outstanding model problems. There are still some things left however:

* The ?PERIOD is currently not handled correctly in eqc_fuse.
* We still need to clean up code which is not part of the system anymore.