Introduce circuit disable/reenable in the model.
The idea is to support manual overrides of fuses by an administrator.
You can put a fuse into a "blown" state which can only be undone by
the administrator. This is useful if your system supports the circuit
breaker directly, as you can use it to carry out maintenance on, say,
your database system.