Provide a timer infrastructure which is precise.
Steal most of the timer code from jlouis/dht and adapt it to this project.
We can now run timing as part of the cluster, and we can slowly move
toward a system where time is mocked in the main system by means
of eqc_mocking rather than our own mock. It makes for clearer code,
but it also makes for simpler, less intrusion into the SUT.

The next steps are to hook up the fuse_eqc model to the timer code
and make all of this work as it should.