Simply fuse edge cases.
Work toward the next goal, which is to handle the case where we are
processing commands. Do not make the distinction between a gradual and
standard fuse when it comes to its current state. This allows standard
and gradual fuses to both be handling different kinds of states when
command processing.